Clean up and restructure Wave addon
Clean up and restructure Wave addon in a saner way. The goal is to have directories "Skyflow", "ExactTarget", "Salesforce" and "Wave" in /src which are as much independent as possible. The registering of services and routes are handled by providers which are specific to each addon. Also there is now a new SalesforceAuthenticator to handle OAuth2 username-password and Web Server Authentication flows details. The app/db.php now contains a database configuration that enables development on localhost. Use dev.sql to create skyflow database and user.