Refactoring and chaining services
There is a new service config parameter "extension" which is the url that is after the endpoint and version. The setters for a service provider, endpoint, version, extension are now protected to ensure they are not hacked from a flow class. Services are now chained. There is a new Service attribute parentService so that the service knows about his parent and can copy its configuration (provider, endpoint, version) ; extension of a parent service is automatically prepended to child service on child service constructor. Services now inherit methods addService() and getService() from the Facade class so we can add child services.