Remove use of non PHP built-in SplEnum in InsightsExternalData domain class and don't inherit AbstractModel
SplEnum is SPL type that is not PHP built-in. It is better not to use it. I also removed inheritance of AbstractModel for InsightsExternalData and InsightsExternalDataPart because of the abstract method getType() which i used to get the database table name in which the model is saved : InsightsExternalData and InsightsExternalDataPart are not meant to be saved in database.