Fill the ParentService for user instantiated services and make service config optional in constructor, refactor WaveExternalDataService::config
Providing the parentService is required for WebServices otherwise the required service endpoint is not copied from parent service to child service and it throws an Exception. The $config service constructor parameter can now be null so that we can instantiate child services without providing any custom config. I also refactored the WaveExternalDataService which does not need its $config parameter anymore.