New service SalesforceQueryService
There is a new SalesforceQueryService to create a query and store the query result. Fieldnames aliases can be set and transforms callbacks can be defined for queried fields. Using the useHeadings option we can specify to use aliases in the records query result or use exclusively the field names as written in the query. Using the includeMissingRelationsFields option we can include related objects queried fields that should normally not be included in the records result because there is no related object associated to the record. A processed query is automatically locked so that we cannot change the FROM, WHERE clauses or add fields, aliases, transforms without the need to create a new query.